% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECHOcovariates-dt.R
\docType{package}
\name{ECHOcovariates-package}
\alias{ECHOcovariates}
\alias{ECHOcovariates-package}
\title{ECHOcovariates: ECHO-study Covariates Workflow}
\description{
Contains ECHO-specific functions for working with covariates/metadata (e.g., ACES, PFAS, PSWQ).
}
\keyword{internal}
