% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_PFAS.R
\name{generate.pfas.data.frame}
\alias{generate.pfas.data.frame}
\title{PFAS dataframe}
\usage{
generate.pfas.data.frame(pfas.dataset.filepath)
}
\arguments{
\item{pfas.dataset.filepath}{file path to PFAS dataset (BLIS)}
}
\value{
a curated data.frame
}
\description{
PFAS dataframe
}
